Do not build classy lens things with haddock - move types and lens to nested modules
